Update GPTerminal.py
works most of the time for simple stuff. it has issues with some of the multi-line stuff still. and still has issues with being given new tasks after completing the first one. the AI sees and responds to it but the checks for pending commands doesn't.

next revision will hopefully be broken out into clean functions with some simple multithreading.